+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 #### :rocket: Enhancement
+* [#26](https://github.com/lblod/frontend-lpdc/pull/26) [LPDC-1359] add creator and lastmodifier user ([@wolfderechter](https://github.com/wolfderechter))
 * [#25](https://github.com/lblod/frontend-lpdc/pull/25) [LPDC-1411] upgrade to appuniversum v3 ([@wolfderechter](https://github.com/wolfderechter))
 * [#12](https://github.com/lblod/frontend-lpdc/pull/12) [LPDC-1412] Rich text editor update ([@Windvis](https://github.com/Windvis))